refactor: move scanlines into internal package
